refactor: replace Simple_Resource_Ability with standalone Style_Best_Practices_Ability
Remove generic abstraction per review feedback. The style best-practices
resource is now a dedicated ability class matching sibling patterns.

Ref: ED-24840
