+++ modules/mcp/module.php
@@ -76,7 +76,7 @@ class Module extends BaseModule {
 		( new Abilities\Create_Page_Ability() )->register();
 		( new Abilities\Get_Globals_Ability() )->register();
 		( new Abilities\Create_Element_Ability() )->register();
-		( new Abilities\Simple_Resource_Ability( 'elementor://style/best-practices', __( 'Style Best Practices', 'elementor' ) ) )->register();
+		( new Abilities\Style_Best_Practices_Ability() )->register();
 	}
 
 	public function register_server( $adapter ) {
@@ -103,7 +103,7 @@ class Module extends BaseModule {
 				'elementor/create-element',
 			],
 			[
-				'elementor/resource-style-best-practices',
+				'elementor/style-best-practices',
 			],
 			[]
 		);
